[Improve](InPredict) function_collection_in handle in predict
